Change memory reading to 64bit
Changed pointers to IntPtr to fit 64bit.
The offsets stay as Int32 because because all of the offsets still fit in Int32 and you can add Int32 to IntPtr easier than Int64.

Also added offsets file for patch 13.7.1.
There are new "Level" and "NameLength" offsets.
The GameObject.Name is now also a packed String but it doesn't exist for some GameObjects (handled by special NameLength check that defaults to an empty name).
Since the biggest offset is now "Level" instead of "EXP" the GameObject buffer size had to be adjusted to "Level" + sizeof(int).